virt-v2v: add FIXME about warm migration
Rich is quite concerned that our current approach to warm migration /
virt-v2v-in-place may result in a bunch of weird behavior during warm
migrations in the future and recommends only using virt-v2v-in-place
with a full xml description of the vm that will be migrated. Add a FIXME
comment so that it doesn't get forgotten.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>