Replace firstboot with new RunOnce task
When running the script during the virt-v2v some of our commands
for example `Get-Disk` returns nothing so we needed some alternative
way to run the scripts.
I have chosen the way to use the `virt-win-reg`
and specify a new `RunOnce` task which will be run by the Windows once
everything is ready. With the `virt-win-reg` was another issue that
we couldn't specify multiple disks to the arguments. So I needed to define
the domain on which we will run the `virt-win-reg`.

So the current workflow now is:
1. Migrate the VM from VMware using `virt-v2v`
2. Define the VM using `virsh define` which will be cleaned up after
   the process finishes
3. Upload our scripts to the VM using `virt-customize`
4. Change the Windows registry using `virt-win-reg`

Signed-off-by: Martin Necas <mnecas@redhat.com>