Rename VSphereUsesEl9VirtV2v() to VSphereColdLocal()
The name of the function no longer makes sense since both of the
different migration paths uses the same el9 image to do the guest
conversion. So rather than naming the function according to what action
should be taken, name it based on what it's actually checking. There
could be a much better name here, but I couldn't come up with one.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>