Update k8s.md
For installing forklift, `kubectl apply` command is provided. The https url in that command is incorrect, it is pointing to old /konveyor/forklift-operator/... repo.
This PR is to modify that url to point to new artifacts from /kubev2v/forklift/.. repo.
Users directly use the `kubectl` command and then the artifacts from the old archived repo gets installed.

Signed-off-by: Niket Chavan <93035123+cniket@users.noreply.github.com>