[MTV-595] Add customize disk image step to virt-v2v pod (#983)
Ref: https://issues.redhat.com/browse/MTV-595

**Issue:**
When migrating a simple RHEL8 VM from VMware 7 to OCPv 4.13 the name of
the network interfaces changes and the static IP configuration for the
VM no longer works

**What this pull request do:**

a. If the `os` is RHEL and user opted to `preserveStaticIPs` and the
source is `vcenter` :

we look for network manager configuration files:
`/etc/sysconfig/network-scripts/`   - for sysconfig scripts
`/etc/NetworkManager/system-connections/` - for network connection files

b. If we find configuration files that can connect `ip address` to
`interface name` we map `interface name` to mac and address.

c. we create a `/etc/udev/rules.d/70-persistent-net.rules` file to pin
the `interface names` to `mac addresses`

**Limitations:**

Only work for network manager
Require the source Virtual machine to run while MTV is collecting
inventory to create the mac to ip mapping
 
**Screenshot**

![screenshot-console-openshift-console apps mtv82 rhos-psi cnv-qe rhood
us-2024 08
27-13_47_24](https://github.com/user-attachments/assets/4563eee3-54fd-4db1-80c3-6bc47eb084d8)

---------

Signed-off-by: yaacov <kobi.zamir@gmail.com>