MTV-1354 | virt-v2v add pod anti affinity to the guest conversion pod
Issue: Right now we leave the placement of the conversion pod up to the
kubernetes scheduler. But it can cause issues in large scale enviroments
as the scheduler can place all migration pods on one node. This can
cause the node to run out of the resources and can slow down the
migration.

Fix: Add pod anti affinity which will distribute the migration pods
evenly across the nodes.

Signed-off-by: Martin Necas <mnecas@redhat.com>