MTV-1366 | Add virt-customize flow to the warm migration
Issue: Right now we have added a new way to inser the scripts into the
guest VM. This is done only on the cold local migration, so the warm
migration or the cold remote migration is completley missing this.

Note: This fix is just for the 2.6.z as in the 2.7.0 we should el9 flow
as proposed in the https://github.com/kubev2v/forklift/pull/996.

This fix include:
- Split the warm migration script into separate functions.
- Add the output from the virt-v2v to the $DOMAIN_DIR so we can get the
  domain xml for the disks and OS.
- Add a virt-customize for windows and rhel both with hardcoded script
  paths for running the scripts and uploading the files. I decided for
the hardcoded paths as this is just quick and temporar solution. We
should come up with more dynamic solution in 2.7.0.
- Link the mounted PVCs to the virt-v2v $DOMAIN_DIR as that is the place
  where the virt-v2v points the disks. Alternativley we could edit the
domain but I did not wanted to go around it and keep the same flow as
the cold migration.

Signed-off-by: Martin Necas <mnecas@redhat.com>