MTV-1366 | Add virt-customize to warm migration and OVA
Issue:
Right now we run the virt-customize only on the cold local migration.
But we would wanted to run it on all vmware migrations.

Fix:
This commit adds a new stage to the converison pod which scans the guest
to get the operating system. It is using the tool `virt-v2v-inspector`
which generates inspection xml file. This file contains lot of
information about guest, including the OS. This tool is run after the
`virt-v2v` step because for the OVA and cold local migration we dont have
the disk available before that step.
Additionally this patch adds a new `inspection` endpoint to the conversion
pod. The controller requests the endpoint to get the operating system so
it could be set as a preference to the kubevirt.

New known issues:
The OVA won't be able to preserve static IP addresses as we get the IP
adress configuration by form the vmware guest agent when the VM is
running. There might be some information inside the ova file but right
now this is not fixed.

Signed-off-by: Martin Necas <mnecas@redhat.com>