MTV-446 | Warm: Add cbt check before migration
Issue:
Right now we have a check to fail the migration when we start to create
the VM which does not have the CBT enabled. We should throw error much
earlier before the customer even start the migration.

Fix:
Add a CBT check to the plan validation. This fix right now affects only
VMware migrations.

Signed-off-by: Martin Necas <mnecas@redhat.com>