run/tests: Move to a dir based approach to scale easier
Before we generated the test data in the test script, but this made it hard
to include new test data.

By moving to a directory based approach we can easily copy n paste /etc/ dirs
from real environments and test them.
With the tradeoff to make it slightly more difficult to manage by hand, and having
more files.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>