vddk validation: don't restart validator pod on failure
If the vddk validator pod fails, we don't need to keep re-trying. The
container simply checks for the existence of a file, so restarting
the pod is unlikely to change anything.

In addition, by specifying `Never` for the restart policy, the completed
pod should be retained for examination after the job fails, which can be
helpful for determining the cause of failure.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>