MTV-1543 | Fix warm migration scheduler
Issue:
When running the warm migration with VMs more than the MaxInFlight disks.
The VMs over this number won't start the migration till the cutover.
Once the cutover is started the VMs which were not started migrate same
as in the old migration. So there is larger downtime.

Fix:
Ignore the cost when in `CopyingPaused`, this allows the other
Migrations to start as the cost is reduced when the VM disk transfer is
finished.

Note:
This patch also improves the cold migration as it ignores the count when
creating the VM so the other VM migrations can get started.
And dramaticially improves the warm migration time as we are not waiting
for the guest conversion as we already have the disk so we start the
guest conversion and do not halt the scheduler.

Fixes: https://issues.redhat.com/browse/MTV-1543

Signed-off-by: Martin Necas <mnecas@redhat.com>