MTV-1537 | Optimalise the plan scheduler (#1088)
Issue: When we start the warm migration with VM which has a lot of disks
we wait for the whole VM to get migrated. We do not ignore the disks
that has been already migrated. This can cause that when we have 2 VMs
with 10 disks each and on each there is one larger disk the whole
scheduler will be halted untill they finish. So even if the left 9 disks
will be done no migration will be started.

Fix: Subtract the finished disks from the disk count.

Fixes: https://issues.redhat.com/browse/MTV-1537

Signed-off-by: Martin Necas <mnecas@redhat.com>