MTV-1477 | Add DataVolume deletion when archiving
Issue:
When running the migration archive we do not delete the DataVolumes.

Fix:
Add deletion of the DataVolumes when running the archive and the VM
migration failed.
At some cases the migration in the context is missing the migration ID
so we can use the plan instead to find the DataVolume.

Fixes: https://issues.redhat.com/browse/MTV-1477

Signed-off-by: Martin Necas <mnecas@redhat.com>