Don't crash when parsing vm yaml from virt-v2v
Make sure to test pointer types for nil before using them. Because yaml
unmarshalling tries to parse as much as it can and ignores extra fields in
both the struct and the data stream, we can end up with a struct that
successfully parses the input stream but is only partially initialized.
Because we're parsing this data from another source, we need to be
defensive and ensure that the data is initialized before using it.

Fixes: https://issues.redhat.com/browse/MTV-1577

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>