MTV-1627 | Fix the PVC search for the progress
Issue:
When monitring the migration process the forklift requests the PVC from
DV ClaimName. On top of that the forklift checks the Annotation "cdi.kubevirt.io/storage.import.importPodName".
But this annotation is missing in the main PVC but is present in the
prime pvc, which format is "prime-$UID_MAINPVC". This annotation points
to the created pod which imports the volume.

Fix:
Search the prime PVC instead of the main PVC, so the forklift is able to
check the importer pod status.

Signed-off-by: Martin Necas <mnecas@redhat.com>