MTV-1536 | Use CDI for disk transfer in cold migration feature toggle
Issue:
This is alternative to https://github.com/kubev2v/forklift/pull/1109
which removes the virt-v2v disk transfer and replaces it with te CDI.
The main dissadvatage of that solution is that in case customer will
have problems with the new CDI method there won't be any easy way how to
go back, to the virt-v2v.

Fix:
This change adds a new feature toggle `feature_vsphere_cold_cdi` in the
forklift controller. By default the feature will be enabled and in case
there will be problems such as disk corruption the user can disable it.

Signed-off-by: Martin Necas <mnecas@redhat.com>