Reorder warm migration flow for vSphere VMs
When performing a warm migration of a vSphere VM, the current version of
Forklift creates the snapshot for each precopy, gets the change ID for the
previous snapshot, and then deletes the previous snapshot. Per the VMware
documentation (see MTV-1656), the correct flow is to record the change
IDs for the old snapshot, delete it, and then create the new snapshot.

This patch implements the correct flow by adding vSphere-specific steps to
the warm migration itinerary to ensure that old snapshots are cleaned
up and change IDs are recorded before creating new snapshots. Due to
the change in flow it's now necessary to store the change IDs somewhere
between itinerary phases, so a list of disk/changeid pairs has been added
to the Precopy struct. This change requires the CRs to be regenerated.

Correcting the warm migration flow appears, at least in some
cases, to address the filesystem corruption documented in
https://issues.redhat.com/browse/MTV-1656.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>