MTV-1656 | Disable snapshot consolidation
Issue: When deleting the snapshot the users with slow storage could
fail to get data corruption due to the consolidation process. The
consolidation writes the snapshot overlay to the disk and if the user has
slow storage it can take a long time. However, the Forklift Controller does not
check if the consolidation is finished and creates a new snapshot. This new
snapshot could be created in the middle of the writing process and it would
not see the changes in the underlying image from the consolidation.

Wrong order
1. Delete snapshot
2. Create snapshot
3. Consolidate

Correct order
1. Delete snapshot
2. Consolidate
3. Create snapshot

Fix: Disable consolidation in the middle of migration and consolidate in
RemoveFinalSnapshot phase. This will make the warm migration faster as
we don't need to wait for the consolidation but at the same time we will
consolidate at the end of migration so we won't cause any issues on the
users env.

Ref: https://issues.redhat.com/browse/MTV-1656

Signed-off-by: Martin Necas <mnecas@redhat.com>