Fix virt-v2v image build
Issue: When running `make build-virt-v2v-image` the build failed due to
some packages throwing 404.

Fix: I needed to change the rpm url from AppStream to BaseOs for perl
packages. It was not detected by the `./hack/virt-v2v-rpm-deps.sh` as
the packages did not change (did not got updated) but only their
location changed. Might be issue in the bazeldnf rpmtree.

Fixes: https://github.com/kubev2v/forklift/pull/1205

Signed-off-by: Martin Necas <mnecas@redhat.com>