MTV-1717 | Wait for DV status
Issue: When CDI is managing lot of DVs at once it can take some time
before the CDI reconciles the DVs status and start the import. In the
function `updateCopyProgress` we check the DV status and if it is paused
we do continue and as next phase we have snapshot removal. So if the DV
is still in paused state and we start the cutover and MTV skips over the
phase thinking it already finished as the DV is in paused.

Fix: Add WaitForDataVolumesStatus phase to wait until the DVs do not have
the paused status.

Ref: https://issues.redhat.com/browse/MTV-1717

https://github.com/kubev2v/forklift/blob/037657463bc0a68ced995c75e92871b929e999d7/pkg/controller/plan/migration.go#L1573-L1575

Signed-off-by: Martin Necas <mnecas@redhat.com>