MTV-1442 | Allow to skip shared disks
Issue: When migrating a VM with shared disks the virt-v2v transfers all
disks that are attached to the VM, that includes the shared disks. But
if we will transfer 2 VMs with the same shared disk the disk will be
migrated twice. This takes a lot of additional resources.

Fix: There are multiple solutions, which vary in complexity. I have
chosen the simplest and fastest solution by adding a new Plan parameter
migrateSharedDisks this determines if the shared disks should be
migrated. By default, it's set to True as that's the default virt-v2v
behaviour. If the toggle is turned off the plan will filter out the
shared disks and use KubeVirt CDI for disk transfer and
virt-v2v-in-place for the guest conversion.

Ref: http://issues.redhat.com/browse/MTV-1442

Signed-off-by: Martin Necas <mnecas@redhat.com>