MTV-1596 | Add new NetworkNameTemplate field to plan CR (#1327)
Ref:
https://issues.redhat.com/browse/MTV-1596

RFE:
Requesting the ability for choosing a custom name for the network and
disks during migration for VM's. Currently migrations end up with
'net-0' or 'vol-1' and cannot be modified. A newly created vm typically
allows for you to name these.

Fix:
Add an optional network template:
```
NetworkNameTemplate is a template for generating network interface names in the target virtual machine.
It follows Go template syntax and has access to the following variables:
   - .NetworkName: If target network is multus, name of the Multus network attachment definition, empty otherwise.
   - .NetworkNamespace: If target network is multus, namespace where the network attachment definition is located.  
   - .NetworkType: type of the network ("Multus" or "Pod")
   - .NetworkIndex: sequential index of the network interface (0-based)
The template can be used to customize network interface names based on target network configuration.
 Note:
   - This template can be overridden at the individual VM level
   - If not specified on VM level and on Plan leverl, default naming conventions will be used
 Examples:
   "net-{{.NetworkIndex}}"
   "{{if eq .NetworkType "pod"}}pod{{else}}multus-{{.NetworkIndex}}{{end}}"
```


In the plan YAML:

```
spec:
   ...
   networkNameTemplate:   "net-{{.NetworkIndex}}"

   vms:
    - name: some-vm
      networkNameTemplate:   "{{if eq .NetworkType "pod"}}pod{{else}}multus-{{.NetworkIndex}}{{end}}"
```

Screenshots:

![plan-yaml-custom-net](https://github.com/user-attachments/assets/8e8de3af-3667-4a0f-82fc-71cdfe3309b8)


![vm-yaml-custom-net](https://github.com/user-attachments/assets/62ffe4bf-078b-4c1e-8972-072b361eddb9)

Signed-off-by: yzamir <kobi.zamir@gmail.com>