MTV-2018 | Preserve static gateway from different subnet
Issue: When running the migration with preserve static IPs, Forklift does not
preserve gateway of a NIC from different subnet.

Fix: Right now we are checking if the gateway is from the subnet as
correlation between the nic and networks. We need to start using
different correlation. The VMware API is exposing the `device` which is
the index of the NICs returned from the API in order.

Ref: https://issues.redhat.com/browse/MTV-2018

Signed-off-by: Martin Necas <mnecas@redhat.com>