MTV-2093 | Sort the disks by bus
Issue: When running a migration of a VM which has various bus
controllers the PVs are not populated by correct disk. This causes that
some larger disks are in small PVs.
This issue happens due to the libvirt domain-VMware API missmatch.
The virt-v2v is outputing the disks in order as they are in the libvirt
domain but Forklift is using the order from API and sorting the disks by
key. The libvirt is sorted in order of devices SCSI, SATA and IDE wheras
sorting by key from VMware API results in order IDE, SATA and SCSI.

Fix: Gather information about the disk busses and sort the disks by
SCSI, SATA and IDE. And within each of these categories we sort by the
key from lower to higer.

Ref: https://issues.redhat.com/browse/MTV-2093

Signed-off-by: Martin Necas <mnecas@redhat.com>