MTV-1966 | Add extra shared disk validations
Issue:
Right now when the user will migrate two VMs with the shared disk and
the migrateSharedDisks enabled we do not give warning to the user that
the disk will be duplicated. Neither if the PVC is already located in
the namepsace.

Fix:
Add extra validation for the VMs to show if there would be some disk
duplication.

Ref: https://issues.redhat.com/browse/MTV-1966

Signed-off-by: Martin Necas <mnecas@redhat.com>