MTV-1796 | Cancel VM only once
Issue: Right now when the user canceles one VM inside the migration the
cancel method gets called many times as the other VMs from the plan are
still running and the plan is being executed. We should skip the
execution of those VMs.

Fix: Check if the VM was marked as completed before executing the cancel
method.

Ref: https://issues.redhat.com/browse/MTV-1796

Signed-off-by: Martin Necas <mnecas@redhat.com>