Fix shared disks validations
Issue: The condition is adding by specific type. If we want to have
different condition msg to different VMs within the same plan we need to
change the type so the condition is not overriden by next condition.
This way all VMs in the items list had message of the last VM.

Fix: Add VM specific condition types.

Ref:
- https://issues.redhat.com/browse/MTV-2206
- https://issues.redhat.com/browse/MTV-2207
- https://issues.redhat.com/browse/MTV-2209

Signed-off-by: Martin Necas <mnecas@redhat.com>