MTV-2199 | Deprecate plan.spec.DiskBus parameter
Issue: The virt-v2v converts the guests to work only with the virtio,
which includes updating the Linux configurations to match the new device
names (vda, vdx...) and for Windows where it edits the boot section to
work with the virtio-blk drivers. The aim of the DiskBus parameter was
to allow the users to specify the Bus because KubeVirt only allows
hotplug disks with the SCSI bus. This issue was moved to the KubeVirt
and will be resolved on their end.

Fix: Deprecate the parameter from 2.8

Ref: https://issues.redhat.com/browse/MTV-2199

Signed-off-by: Martin Necas <mnecas@redhat.com>