MTV-2236 | Use unshare seccomp
Issue: Right now the migration using the quay.io/kubev2v/forklift-virt-v2v:latest
fails due to the passt missing permissions. The root cause of this issue is in the
passt using the `unshare(2)` syscall. The passt was added to the virt-v2v libguestfs
which introduced this issue. The CRI-O is missing the `unshare(2)` from the seccomp settings.
We need to specify to use the seccomp which allows the `unshare(2)` syscall and add it
to the MTVs SecurityContextConstraints.

Fix: Add the unshare seccomp profile to the guest conversion pod.

Ref: https://issues.redhat.com/browse/MTV-2236

Signed-off-by: Martin Necas <mnecas@redhat.com>